+++ cms/internal/pkg/models/location.go
@@ -6,8 +6,8 @@ type Location struct {
 	ID        int
 	Name      string
 	Slug      string
-	Latitude  float32
-	Longitude float32
+	Latitude  float64
+	Longitude float64
 
 	CreatedAt time.Time
 	UpdatedAt time.Time
